Make live transactions clone-free and locally relowered
