--- .github/workflows/b2-transaction-integration.yml
@@ -1,47 +1,0 @@
-name: B2 transaction integration
-
-on:
-  push:
-    branches:
-      - arch/b2-incremental-runtime
-
-permissions:
-  contents: write
-
-concurrency:
-  group: b2-transaction-integration
-  cancel-in-progress: true
-
-jobs:
-  integrate:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: arch/b2-incremental-runtime
-          fetch-depth: 1
-
-      - name: Apply clone-free local transaction migration
-        run: |
-          python3 scripts/b2_transaction_helper_fix.py
-          python3 scripts/b2_transaction_integration.py
-
-      - name: Format and validate B2
-        run: |
-          cargo fmt --all
-          cargo clippy -p noon-core -p noon-compile -p noon-runtime -p noon-render-wgpu -p noon-web --all-targets --all-features -- -D warnings
-          cargo test -p noon-core --all-features
-          cargo test -p noon-compile --all-features
-          cargo test -p noon-runtime --all-features
-          cargo test -p noon-render-wgpu --all-features
-          cargo test -p noon-web --all-features
-
-      - name: Commit integrated source and remove helpers
-        run: |
-          rm .github/workflows/b2-transaction-integration.yml scripts/b2_transaction_integration.py scripts/b2_transaction_helper_fix.py
-          git config user.name "Yongkyun Shin"
-          git config user.email "yongkyuns@gmail.com"
-          git add -A
-          git commit -m "Make live transactions clone-free and locally relowered"
-          git push origin HEAD:arch/b2-incremental-runtime
